Safari-609.1.11.1/LayoutTests/media/modern-media-controls/time-label/time-label-white-space-nowrap-expected.txt

Testing that time labels inside media controls have 'white-space: nowrap'.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.internals.shadowRoot(media).querySelector('.time-label') instanceof HTMLDivElement became true
PASS timeLabel.ownerDocument.defaultView.getComputedStyle(timeLabel).whiteSpace is "nowrap"

PASS successfullyParsed is true

TEST COMPLETE
